install chocolatety directly instead of winget
I was having trouble getting variables to set and expose properly to be able to start using the "choco" command immediately after installing, with no user input. I tried various ways of setting path variables and/or calling "refreshenv" to no avail, but this "iex" call, combined with modifying path directly, worked and I was able to use "choco" in the script called from this script.