fix snippet autocomplete by configuring nvim cmp earlier
Snippets were appearing in the completion dropdown just fine on my previous machine, but weren't on my new machine with the same neovim config (and the same environment as far as I could tell other than that was Win10 and this is Win11). Maybe some sort of race condition?
idk, but this works fine for now.